+++ CHANGELOG.md
@@ -6,6 +6,7 @@ HEAD
 
 * Fixed `-Wparentheses` warning introduced in v5.6.5 (PR #335 by @nuket)
 * Added `.mbedignore` for ARM mbdeb (PR #334 by @nuket)
+* Fixed  `JsonVariant::success()` which didn't propagate `JsonArray::success()` nor `JsonObject::success()` (issue #342).
 
 v5.6.5
 ------